Add `Maybe [DTyVarBndr]` fields to `DInstanceD`/`DStandaloneDerivD` (#117)
This resolves #117 by implementing the trick described in that issue.
Alas, this trick does not work on all versions of GHC that
`th-desugar` currently supports, so some CPP hackery is employed to
produce slightly more sensible error messages in the event that the
trick is used on incompatible GHC versions.