Fix #123 by removing SYB code from hot loops (#124)
`L.H.T.Desugar.Expand` was using code from `syb` in a frequently
accessed loop (`expand_con.go`), which is the immediate cause of
the slowness observed in #123. This patch replaces this `syb` code
with a more verbose—but much more efficient—hand-written
traversal. The net benefit of doing this is that the program in #123
now compiles in less than 1 second.

While I was in town, I noticed that there were some places in
`th-desugar` that called `geq` when a simple `(==)` would have
sufficed. This requires deriving `Eq` instances for various data
types, but given that `syb` has been observed to introduce slowness,
I think these extra instances are justified.

Fixes #123.