Remove catch-all pattern in decToTH
This tripped me up recently when I tried experimenting with
GHC 8.9's new `KiSigD` constructor. I forgot to update `decToTH`
because it failed to warn about a missing case for `KiSigD` due
to the catch-all wildcard pattern. This patch changes `decToTH`
to explicitly match on every `Dec` constructor to avoid this
pitfall in the future.