Fix #100 by making maybeReifyCon/findRecSelector GADT-aware
Deep in the swamps of `L.H.TH.Desugar.Reify` is the `maybeReifyCon`
function, which locally reifies the type of a data constructor. This
code was deeply buggy in the presence of GADT constructors, as
`maybeReifyCon` would copy information from the data type head
instead of the data type in the constructor's result type. This led
to a variety of bugs observed in #100.

This fixes the issue by giving `maybeReifyCon` a slightly different
code path for GADT constructors. See the internal `con_to_type`
function for the full story.

A similar fix was also applied to `findRecSelector`, which locally
reifies the type of a record selector. This also needs slightly
different logic for GADTs, so I crafted a custom `RecSelInfo` data
type for this purpose to encode the extra information that GADTs
provide (in the form of their return types).

Fixes #100.