Fix #128 by using the right Name in find_assoc_type_kind
I accidentally used the `Name` of the parent class, not the type
variable name, when looking up entries in the mapping from class type
variable names to their kinds in
`L.H.TH.Desugar.Reify.find_assoc_type_kind`, which caused #128. Oops.

It turns out that the existing code which tested `dsReifyType` on an
associated type family wasn't interesting enough to catch this
mistake, so I spruced up the existing code so that it tickles this
particular edge case.