Remove conExistentialTvbs, simplify spec of getRecordSelectors
As observed in https://github.com/goldfirere/singletons/issues/443,
the `conExistentialTvbs` function is completely broken. While it may
techincally be possible to repair `conExistentialTvbs` (see
https://github.com/goldfirere/singletons/issues/443#issuecomment-619578414),
it would require adopting _ad hoc_ heuristics that are difficult to
specify.

I decided to take a step back and ask: why does `conExistentialTvbs`
even exist in the first place? In the case of `th-desugar`, its only
call site is `getRecordSelectors`, which uses `conExistentialTvbs` to
filter out naughty record selectors. But this is a heavy price to pay
for the convenience of supporting this corner case. Moreover, we're
`th-desugar` is not internally consistent about supporting this
corner case. See `L.H.T.Desugar.Reify.maybeReifyCon`, which states
that it does not "ferret out" naughty record selectors.

Bottom line: let's not bother trying to compute existential type
variables, especially since they're so much trouble. This means that:

1. `getRecordSelectors` will include naughty record selectors in its
   results. This is a mite unfortunate, but I would prefer to wait
   for someone to shout about this before investing more effort into
   it.
2. Some validity checks in `singletons` which use
   `conExistentialTvbs` will now be less precise. This isn't a huge
   deal—it just means that erroneous programs which `singletons`
   used to give specialized error messages for will now just go
   through, resulting in an error message on GHC's behalf.

One knock-on change from this patch is that the first argument to
`getRecordSelectors` is no longer required, as its sole purpose was
to help compute existential type variables.