Require GHC 8.0 or later
This requires GHC 8.0 as the minimum, removing support for GHC 7.8 and 7.10
in the process. As a consequence, we can now remove enormous chunks of unneeded
CPP.

Now that we no longer use `template-haskell`'s `Strict` type, the
`strictToBang` function no longer serves a useful purpose. I've opted
to remove it, which makes this commit a breaking change (albeit a pretty
obscure one).