Don't apply type-variable hack when expanding closed type families
`th-desugar`'s type expansion machinery contains a workaround for
https://gitlab.haskell.org/ghc/ghc/-/issues/9262 that prevents `reifyInstances`
from failing when the argument types contain type variables. This workaround
was applied to both open and closed type families alike. While open type
families certainly need the workaround, as their instances must be queried
using `reifyInstances`, the workaround isn't necessary at all for closed type
families, where all of the equations are immediately accessible.

This patch removes the workaround for closed type families, thereby
fixing #161. The fix itself is a one-line change: just delete the `args_ok`
check in the `DClosedTypeFamilyD` case of `expand_con.go`. This patch looks
more involved than it actually is because `args_ok` and friends now must also
be surrounded with CPP to avoid GHC warning that it is unused on modern GHCs.
While I was in town, I added some comments to `args_ok` to explain the
workaround and make it more obvious why it exists.