Use `DeriveLift` instead of TH to define `Lift` instances
Now that we depend on GHC 8.0 as the minimum, we can use `DeriveLift` to derive
all of our `Lift` instances instead of having to define them as orphan instances
using Template Haskell.

This requires `th-abstraction-0.5.*` or later, as that is the first
`th-abstraction` version to define a `Lift` instance for its backported
`Specificity` data type.

Fixes #176.