Make `reifies local definition` test output easier to read
Previously, the test suite was printing the index of each declaration in
`reifyDecsNames`. This led to confusing output, however, as the names in this
list are also numeric, but they usually do not line up with the index numbers
in the list due to the extensive use of CPP.

I have tweaked the test output to simply display the name of the declaration
instead, which is much more straightforward.