Support invisible binders in type-level declarations
Among other changes, there is now a `DTyVarBndrVis` synonym, similar to the
existing `DTyVarBndrSpec` and `DTyVarBndrUnit` synonyms, and various types in
`Language.Haskell.TH.Desugar.AST` are now defined in terms of `DTyVarBndrVis`
instead of `DTyVarBndrUnit`. See the `CHANGES` file for the full details.

This bumps the lower version bounds on `th-abstraction` to `>=0.6`, as this
patch makes critical use of several compatibility shims introduced in that
version of `th-abstraction`.

Fixes #186.