Partial support embedded types in expressions and patterns
This is a partial fix for #204 which adds `DTypeE` and `DTypeP` constructs to
the `th-desugar` AST, which behave exactly like their `TypeE` and `TypeP`
counterparts in the `template-haskell` AST. This is only a partial fix,
however, because `DTypeP` is currently only supported in the clauses of
function declarations. In particular, `DTypeP` is _not_ supported in lambda
expressions, `\case` expressions, or `\cases` expressions. See the "Known
limitations" section of the `README` for more.

In order to add full support for `DTypeP`, we would first need to implement the
design proposed in #210. Until then, this is good enough.