Make unboxed{Tuple,Sum}NameDegree_maybe work with GHC 9.10
This patch changes `unboxedTupleNameDegree_maybe` to handle non-punning tuple
names introduced in GHC 9.10, such as `Unit#`, `Solo#`, `Tuple2#`, `Tuple3#`,
etc. This is because `''(##)` is now syntactic sugar for `''Unit#`, `''(#,#)`
is now syntactic sugar for `Tuple2#`, and so on, so it is important to be able
to recognize both forms.

Similarly, this patch changes `unboxedSumNameDegree_maybe` to handle `Sum2#`,
`Sum3#`, etc.

Fixes #213.