Remove outdated Note and code after DLamCasesE changes in #218
After the `DLamCasesE`-related changes in #218, most of the material in `Note
[Auxiliary tuples in pattern matching]` is out of date, as we no longer use
unboxed tuples as intermediate data types when desugaring pattern matches.
Moreover, several functions which construct these unboxed tuples are no longer
used anywhere in `th-desugar`. This patch removes both the Note and unused
functions.

The Note included references to parts of the code where intermediate (but not
unboxed) tuples are still used in the `th-desugar` code base, which are still
relevant. As such, this patch moves these references to another part of the
code.