dMatchUpSAKWithDecl: Fix treatment of invisible foralls without corresponding @-binders
After this patch, `dMatchUpSAKWithDecl` is now able to handle a standalone kind
signature with an invisible `forall` without a corresponding `@`-binder, as
seen in the example below:

```hs
type D :: forall (a :: Type). Type
data D
```

While I was in town, I did some code cleanup in `dMatchUpSAKWithDecl` to make
sure that all of the various error conditions are checked more thoroughly,
documenting what parts of the code rely on which preconditions more precisely.
I also applied the same treatment to `matchUpSAKWithDecl`.

Fixes #228.